Document Q&A with Gemma/requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain-community
python-dotenv
pypdf